+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 7.4.0 IN PROGRESS
+  * ERM-4126: Publish Kafka DELETE events for agreement lines (entitlements)
 
 ## 7.3.2 2026-06-04
   * ERM-4087: Agreements: Bump access-control-grails version to handle jackson vulnerability